+++ Models/AwardHistory.cs
@@ -22,7 +22,7 @@ namespace xcart.Models
 
 
         [ForeignKey("EmployeeId")]
-        public int EmployeeId { get; set; }
+        public long EmployeeId { get; set; }
         public virtual User Employee { get; set; }
 
         [ForeignKey("AwardId")]
@@ -30,7 +30,7 @@ namespace xcart.Models
         public virtual Award Award { get; set; }
 
         [ForeignKey("PresenteeId")]
-        public int? PresenteeId { get; set; }
+        public long? PresenteeId { get; set; }
         public virtual User Presentee  { get; set; }
 
         [Column("Point")]
